Fix for Moodle validation issue 6

+++ domoscio/linkto.php
@@ -76,23 +76,43 @@ if (has_capability('moodle/course:create', $context)) {
         echo html_writer::tag('div', get_string('linkto_intro', 'domoscio').html_writer::tag('b', $linked_module->display." - ".$notion->name, array('class' => '')), array('class' => 'well'));
         echo html_writer::link("$CFG->wwwroot/mod/domoscio/view.php?id=$cm->id", '<< '.get_string('back_btn', 'domoscio')."&nbsp");
 
-        $quizzes = $DB->get_records('quiz', array('course' => $course->id), '', 'id,name');
+        $quizzes = $DB->get_records('quiz', array(), '', 'id,name');
         $scorms = $DB->get_records('scorm', array(), '', '*');
 
         $list = '';
 
         foreach($quizzes as $quiz)
         {
-            $icon = html_writer::tag('img', '', array('src'=>$OUTPUT->pix_url('icon','quiz','quiz',array('class'=>'icon')), 'class'=>'activityicon', 'alt'=>'disable'));
-            $url = html_writer::link("$CFG->wwwroot/mod/domoscio/linkto.php?id=$cm->id&notion=$kn&exo=quiz_$quiz->id", $icon." ".$quiz->name);
-            $list .= html_writer::tag('h5', $url, array('class' => 'well well-small'));
+            $quiz_cm = $DB->get_record('course_modules', array('instance' => $quiz->id, 'module' => 16), '*');
+
+            if($quiz_cm)
+            {
+                $quiz_context = context_module::instance($quiz_cm->id);
+            }
+
+            if(has_capability('mod/quiz:manage', $quiz_context))
+            {
+                $icon = html_writer::tag('img', '', array('src'=>$OUTPUT->pix_url('icon','quiz','quiz',array('class'=>'icon')), 'class'=>'activityicon', 'alt'=>'disable'));
+                $url = html_writer::link("$CFG->wwwroot/mod/domoscio/linkto.php?id=$cm->id&notion=$kn&exo=quiz_$quiz->id", $icon." ".$quiz->name);
+                $list .= html_writer::tag('h5', $url, array('class' => 'well well-small'));
+            }
         }
 
         foreach($scorms as $scorm)
         {
-            $icon = html_writer::tag('img', '', array('src'=>$OUTPUT->pix_url('icon','scorm','scorm',array('class'=>'icon')), 'class'=>'activityicon', 'alt'=>'disable'));
-            $url = html_writer::link("$CFG->wwwroot/mod/domoscio/linkto.php?id=$cm->id&notion=$kn&exo=scorm_$scorm->id", $icon." ".$scorm->name);
-            $list .= html_writer::tag('h5', $url, array('class' => 'well well-small'));
+            $scorm_cm = $DB->get_record('course_modules', array('instance' => $scorm->id, 'module' => 18));
+
+            if($scorm_cm)
+            {
+                $scorm_context = context_module::instance($scorm_cm->id);
+            }
+
+            if(has_capability('mod/scorm:viewreport', $scorm_context))
+            {
+                $icon = html_writer::tag('img', '', array('src'=>$OUTPUT->pix_url('icon','scorm','scorm',array('class'=>'icon')), 'class'=>'activityicon', 'alt'=>'disable'));
+                $url = html_writer::link("$CFG->wwwroot/mod/domoscio/linkto.php?id=$cm->id&notion=$kn&exo=scorm_$scorm->id", $icon." ".$scorm->name);
+                $list .= html_writer::tag('h5', $url, array('class' => 'well well-small'));
+            }
         }
         $icon = html_writer::empty_tag('img', array('src'=>$OUTPUT->pix_url('i/edit'), 'alt'=>get_string('edit'), 'class'=>'smallicon'));
         $url = html_writer::link("#", $icon." ".get_string('create_q', 'domoscio')." (Coming Soon)", array('disabled' => 'disabled'));
@@ -104,52 +124,74 @@ if (has_capability('moodle/course:create', $context)) {
     {
         $selected = explode('_', $exo);
 
-        $mform = new mod_domoscio_linkto_form("$CFG->wwwroot/mod/domoscio/linkto.php?id=$cm->id&notion=$kn&exo=$exo", array('kn_id' => $kn, 'module' => $selected[0], 'exo_id' => $selected[1]));
+        if($selected[0] == 'quiz')
+        {
+            $cap = 'mod/quiz:manage';
+            $module = 16;
+        }
+        elseif($selected[0] == 'scorm')
+        {
+            $cap = 'mod/scorm:viewreport';
+            $module = 18;
+        }
 
-        if ($mform->is_cancelled()) {
+        $cmid = $DB->get_record('course_modules', array('instance' => $selected[1], 'module' => $module));
+        $cm_context = context_module::instance($cmid->id);
 
-          redirect("$CFG->wwwroot/mod/domoscio/linkto.php?id=$cm->id&notion=$kn");
-          exit;
+        if(has_capability($cap, $cm_context))
+        {
+            $mform = new mod_domoscio_linkto_form("$CFG->wwwroot/mod/domoscio/linkto.php?id=$cm->id&notion=$kn&exo=$exo", array('kn_id' => $kn, 'module' => $selected[0], 'exo_id' => $selected[1]));
 
-        } else if ($fromform = $mform->get_data()) {
+            if ($mform->is_cancelled()) {
 
-            foreach($fromform as $k => $value)
-            {
-                if(is_numeric($k))
+              redirect("$CFG->wwwroot/mod/domoscio/linkto.php?id=$cm->id&notion=$kn");
+              exit;
+
+            } else if ($fromform = $mform->get_data()) {
+
+                foreach($fromform as $k => $value)
                 {
-                    $check = $DB->get_record_sql("SELECT *
-                                                    FROM {knowledge_node_questions}
-                                                   WHERE `question_id` = $k
-                                                     AND knowledge_node = $notion->id");
+                    if(is_numeric($k))
+                    {
+                        $check = $DB->get_record_sql("SELECT *
+                                                        FROM {knowledge_node_questions}
+                                                       WHERE `question_id` = $k
+                                                         AND knowledge_node = $notion->id");
 
-                    if($value == 1)
-                    {
-                        if($check == null)
+                        if($value == 1)
                         {
-                            $entry = new stdClass;
-                            $entry->instance = $domoscio->id;
-                            $entry->knowledge_node = $kn;
-                            $entry->question_id = $k;
-                            $entry->type = $selected[0];
-                            $write = $DB->insert_record('knowledge_node_questions', $entry);
+                            if($check == null)
+                            {
+                                $entry = new stdClass;
+                                $entry->instance = $domoscio->id;
+                                $entry->knowledge_node = $kn;
+                                $entry->question_id = $k;
+                                $entry->type = $selected[0];
+                                $write = $DB->insert_record('knowledge_node_questions', $entry);
+                            }
                         }
-                    }
-                    elseif($value == 0)
-                    {
-                        if(!empty($check))
+                        elseif($value == 0)
                         {
-                            $DB->delete_records('knowledge_node_questions', array('question_id' => $k, 'knowledge_node' => $notion->id));
+                            if(!empty($check))
+                            {
+                                $DB->delete_records('knowledge_node_questions', array('question_id' => $k, 'knowledge_node' => $notion->id));
+                            }
                         }
                     }
                 }
+
+                echo get_string('upd_qlist', 'domoscio')."<hr/>";
+                echo html_writer::tag('button', get_string('next_btn', 'domoscio'), array('type' => 'button','onclick'=>"javascript:location.href='$CFG->wwwroot/mod/domoscio/view.php?id=$cm->id'"));
+
+            } else {
+                $mform->display();
             }
-
-            echo get_string('upd_qlist', 'domoscio')."<hr/>";
-            echo html_writer::tag('button', get_string('next_btn', 'domoscio'), array('type' => 'button','onclick'=>"javascript:location.href='$CFG->wwwroot/mod/domoscio/view.php?id=$cm->id'"));
-
-        } else {
-            $mform->display();
         }
+        else
+        {
+            echo get_string('nocapabilitytousethisservice', 'error');
+        }
+
     }
 }
